Return proper HTTP status code and fixed deprection

+++ src/Controller/Frontend/Basket/Base.php
@@ -40,7 +40,7 @@ abstract class Base extends \Aimeos\Controller\Frontend\Base implements Iface
 		{
 			$amount = $attr->getValue();
 
-			if( preg_match( '/^[0-9]*(\.[0-9]+)?$/', $amount ) !== 1 || ( (double) $amount ) < 0.01 )
+			if( preg_match( '/^[0-9]*(\.[0-9]+)?$/', $amount ) !== 1 || ( (float) $amount ) < 0.01 )
 			{
 				$msg = $context->translate( 'controller/frontend', 'Invalid price value "%1$s"' );
 				throw new \Aimeos\Controller\Frontend\Basket\Exception( sprintf( $msg, $amount ) );
